Build on 64-bit architectures, where pointers are larger than ints.
git-svn-id: file:///home/apenwarr/alumnit-svn/public/releases/freesw@10924 6619c942-59fa-0310-bf43-a07aaa81d9ac

+++ utils/wvhash.cc
@@ -30,6 +30,6 @@ unsigned WvHash(const int &i)
 
 unsigned WvHash(const void *p)
 {
-    return reinterpret_cast<unsigned>(p);
+    return reinterpret_cast<unsigned long>(p);
 }
